Move kicad stuff & fix example more
