+++ faebryk/library/__init__.py
@@ -3,5 +3,6 @@
 
 
 import faebryk.library.core
+import faebryk.library.kicad
 import faebryk.library.library
 import faebryk.library.traits
