+++ faebryk/library/library/__init__.py
@@ -4,6 +4,5 @@
 import faebryk.library.library.components
 import faebryk.library.library.footprints
 import faebryk.library.library.interfaces
-import faebryk.library.library.kicad
 import faebryk.library.library.links
 import faebryk.library.library.parameters
