+++ samples/iterative_design_nand.py
@@ -11,7 +11,7 @@ def run_experiment():
     from faebryk.library.library.interfaces import Electrical, Power
     from faebryk.library.library.parameters import TBD, Constant
     from faebryk.library.traits.component import has_defined_footprint, has_defined_footprint_pinmap, has_symmetric_footprint_pinmap, has_interfaces, has_defined_interfaces
-    from faebryk.library.traits.footprint import has_kicad_manual_footprint
+    from faebryk.library.kicad import has_kicad_manual_footprint
 
     # levels
     high = Electrical()
