+++ test/exporters/netlist/kicad/test_netlist_kicad.py
@@ -11,9 +11,9 @@ def test_netlist_graph():
     from faebryk.library.library.interfaces import Electrical
     from faebryk.library.traits.component import has_defined_footprint
     from faebryk.library.core import Footprint
-    from faebryk.library.library.kicad import has_defined_kicad_ref
+    from faebryk.library.kicad import has_defined_kicad_ref
     from faebryk.library.traits.component import has_defined_footprint_pinmap, has_defined_type_description, has_interfaces, has_interfaces_list, has_type_description
-    from faebryk.library.traits.footprint import has_kicad_manual_footprint
+    from faebryk.library.kicad import has_kicad_manual_footprint
     from faebryk.exporters.netlist.graph import make_t1_netlist_from_graph, make_graph_from_components
